Struct mitochondria::move_::core::MoveCell

pub struct MoveCell<T> {
    value: UnsafeCell<T>,
}

A mutable memory location that steals ownership.

Fields
value: UnsafeCell<T>
Methods

impl<T> MoveCell<T>

fn new(value: T) -> Self

Creates a new MoveCell containing value.

Examples

use mitochondria::MoveCell;

let c = MoveCell::new("Hello cytosol!".to_owned());

fn into_inner(self) -> T

Consumes the MoveCell, returning the wrapped value.

Examples

use mitochondria::MoveCell;

let c = MoveCell::new("Hello matrix!".to_owned());

let greeting = c.into_inner();

fn replace(&self, value: T) -> T

Replaces the value of this cell, returning the old one.

Examples

use mitochondria::MoveCell;

let c = MoveCell::new("Hello lysosome!".to_owned());

let greeting = c.replace("Goodbye!".to_owned());

fn as_ptr(&self) -> *mut T

Returns a raw pointer to the underlying value in this cell.

Examples

use mitochondria::MoveCell;

let c = MoveCell::new("Hello centrosome!".to_owned());

let ptr = c.as_ptr();

fn as_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying value.

This call borrows MoveCell mutably (at compile-time) which guarantees that we possess the only reference.

use mitochondria::MoveCell;

let mut c = MoveCell::new("Porins,".to_owned());
*c.as_mut() += " unite!";

assert_eq!(c.into_inner(), "Porins, unite!");